Refactor renderList
